Fix some bugs.
