+++ src/codeu/chat/server/LocalFile.java
@@ -2,6 +2,7 @@ package codeu.chat.server;
 
 import java.io.File;
 import java.io.IOException;
+import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.FileInputStream;
 
@@ -17,9 +18,17 @@ import codeu.chat.common.Message;
 import codeu.chat.common.User;
 import codeu.chat.util.Logger;
 import codeu.chat.util.Uuid;
+import codeu.chat.util.Serializer;
+import codeu.chat.util.Serializers;
 
 public class LocalFile
 {
+    private final static Logger.Log LOG = Logger.newLog(LocalFile.class);
+
+    public static final String MESSAGE_FILE_NAME = "/msgDat.sav";
+    public static final String USER_FILE_NAME = "/usrDat.sav";
+    public static final String CONVERSATION_FILE_NAME = "/cvrsDat.sav";
+
     //Instance varibles for saving the current data of server.
     private final LinkedHashSet<User> users;
     private final LinkedHashSet<ConversationHeader> conversationHeaders;
@@ -35,54 +44,74 @@ public class LocalFile
     //true: All new data been added to this instance should mark the status as modified.
     private boolean isInitialized = false;
 
+    private final Serializer<Collection<Message>> localMessages = Serializers.collection(Message.SERIALIZER);
+    private final Serializer<Collection<ConversationHeader>> localConversationHeaders = Serializers.collection(ConversationHeader.SERIALIZER);
+    private final Serializer<Collection<User>> localUsers = Serializers.collection(User.SERIALIZER);
+
+    private final File userFile;
+    private final File conversationFile;
+    private final File messageFile;
     public LocalFile (File file)
     {
         this.file = file;
         users = new LinkedHashSet<>();
         conversationHeaders= new LinkedHashSet<>();
         messages= new LinkedHashSet<>();
+
+        userFile = new File(file.getPath() + USER_FILE_NAME);
+        conversationFile = new File(file.getPath() + CONVERSATION_FILE_NAME);
+        messageFile = new File(file.getPath() + MESSAGE_FILE_NAME);
+        try
+        {
+            if(!userFile.exists())
+            {
+                userFile.createNewFile();
+            }      
+            if(!conversationFile.exists())
+            {
+                conversationFile.createNewFile();
+            }
+            if(!messageFile.exists())
+            {
+                messageFile.createNewFile();
+            }
+        }
+        catch(IOException exception)
+        {
+            LOG.error("Failed to create new file!");
+            exception.printStackTrace();
+            throw new RuntimeException();
+        }
     }
 
     /**
      * Get a copy of users
      * 
-     * @param   boolean Is the return value going to be saved? If true, the status should be updated.
      * @return  LinkedHashSet<User> Current users from this instance
      */
-    public LinkedHashSet<User> getCopyOfUsers(boolean willBeSaved)
+    public LinkedHashSet<User> getCopyOfUsers()
     {
-        if(willBeSaved)
-        {
-            hasUserModified = false;
-        }
+        hasUserModified = false;
         return new LinkedHashSet<>(users);
     }
     /**
      * Get a copy of conversations
      * 
-     * @param   boolean Is the return value going to be saved? If true, the status should be updated.
      * @return  LinkedHashSet<ConversationHeader> Current conversations from this instance
      */
-    public LinkedHashSet<ConversationHeader> getCopyOfConversationHeaders(boolean willBeSaved)
+    public LinkedHashSet<ConversationHeader> getCopyOfConversationHeaders()
     {
-        if(willBeSaved)
-        {
-            hasConversationModified = false;
-        }
+        hasConversationModified = false;
         return new LinkedHashSet<>(conversationHeaders);
     }
     /**
      * Get a copy of messages
      * 
-     * @param   boolean Is the return value going to be saved? If true, the status should be updated.
      * @return  LinkedHashSet<Message> Current messages from this instance
      */
-    public LinkedHashSet<Message> getCopyOfMessages(boolean willBeSaved)
+    public LinkedHashSet<Message> getCopyOfMessages()
     {
-        if(willBeSaved)
-        {
-            hasMessageModified = false;
-        }
+        hasMessageModified = false;
         return new LinkedHashSet<>(messages);
     }
     /**
@@ -133,39 +162,7 @@ public class LocalFile
     {
         return file.getPath();
     }
-    /**
-     * Check if the User data has been modified.
-     * If this instance is not initialized, it should always return false.
-     * 
-     * @return boolean Status of the User data
-     */
 
-    public boolean hasUserModified()
-    {
-        return hasUserModified;
-    }
-
-    /**
-     * Check if the ConversationHeader data has been modified.
-     * If this instance is not initialized, it should always return false.
-     * 
-     * @return boolean Status of the ConversationHeader data
-     */
-    public boolean hasConversationModified()
-    {
-        return hasConversationModified;
-    }
-    /**
-     * Check if the Message data has been modified.
-     * If this instance is not initialized, it should always return false.
-     * 
-     * @return boolean Status of the Message data
-     */
-
-    public boolean hasMessageModified()
-    {
-        return hasMessageModified;
-    }
     /**
      * The initialization of this instance is complete.
      * Then all new data been added to this instance should mark the status as modified.
@@ -174,4 +171,117 @@ public class LocalFile
     {
         isInitialized = true;
     }
+
+    /**
+   * Save user data
+   * @throws IOException
+   */
+  private FileOutputStream saveUsers() throws IOException
+  {
+    FileOutputStream userStream = null;
+    try
+    {
+      userStream = new FileOutputStream(userFile);
+      localUsers.write(userStream, users);
+    }
+    catch(FileNotFoundException exception)
+    {
+      System.out.println("ERROR:Unacceptable file path");
+      exception.printStackTrace();
+      throw exception;
+    }
+    catch(IOException exception)
+    {
+      System.out.println("ERROR:Failed to get ConversationHeaderStream!");
+      exception.printStackTrace();
+      throw exception;
+    }
+    return userStream;
+  }
+  /**
+   * Save conversation data
+   * @throws IOException
+   */
+  private FileOutputStream saveConversationHeaders() throws IOException
+  {
+    FileOutputStream conversationStream = null;
+    try
+    {
+      conversationStream = new FileOutputStream(conversationFile);
+      localConversationHeaders.write(conversationStream, conversationHeaders);
+    }
+    catch (FileNotFoundException exception)
+    {
+      System.out.println("ERROR:Unacceptable file path");
+      exception.printStackTrace();
+      throw exception;
+    }
+    catch(IOException exception)
+    {
+      System.out.println("ERROR:Failed to get ConversationHeaderStream!");
+      exception.printStackTrace();
+      throw exception;
+    }
+    return conversationStream;
+  }
+  /**
+   * Save message data
+   * @throws IOException
+   */
+  private FileOutputStream saveMessages() throws IOException
+  {
+    FileOutputStream messageStream = null;
+    try
+    {
+      messageStream = new FileOutputStream(messageFile);
+      localMessages.write(messageStream, messages);
+    }
+    catch (FileNotFoundException exception)
+    {
+      System.out.println("ERROR:Unacceptable file path");
+      exception.printStackTrace();
+      throw exception;
+    }
+    catch (IOException exception)
+    {
+      System.out.println("ERROR:Failed to get ConversationHeaderStream!");
+      exception.printStackTrace();
+      throw exception;
+    }
+    return messageStream;
+  }
+  /**
+   * Save all data
+   * @throws IOException
+   */
+  public void saveData() throws IOException
+  {
+    try
+    {
+      if(hasConversationModified)
+      {
+        saveConversationHeaders();
+        LOG.info("Conversation data Saved!");
+        hasConversationModified = false;
+      }
+      if(hasMessageModified)
+      {
+        saveMessages();
+        LOG.info("Message data Saved!");
+        hasMessageModified = false;
+      }
+      if(hasUserModified)
+      {
+        saveUsers();
+        LOG.info("User data Saved!");
+        hasUserModified = false;
+      }
+    }
+    catch(IOException exception)
+    {
+      System.out.println("ERROR:Failed to save data!");
+      exception.printStackTrace();
+      throw exception;
+    }
+  }
 }
